requirements.txt
pyyaml>=6.0
requests>=2.31.0
tabulate>=0.9.0
colorama>=0.4.6
python-dotenv>=1.0.0
